Now fully prepared, but, still not working :)

+++ tests/basic-test.cpp
@@ -20,7 +20,8 @@ public:
     TestActionA() : Action("TestActionA") {}
     
     Result perform(SessionPtr session) const override{
-        std::cout << " Test Action A" << std::endl;
+        BOOST_LOG_SEV(logger, Info) ;
+        // fingerprint(session) << " Hello from Action Test A";
         return done();
     }
     
@@ -28,20 +29,32 @@ public:
         
 boost::shared_ptr<boost::asio::io_service::work> work;
 boost::shared_ptr<boost::asio::io_service> service;
+TimedPtr t;
         
-void terminate(const boost::system::error_code & ) {
+void terminate(const boost::system::error_code & ec ) {
+    if(not ec) {
+    GLOB_LOGGER("general");
+    BOOST_LOG_SEV(logger, Info) << " Shutting down !";
     ControllerManager::getInstance()->terminate();
     work.reset();
+    } else {
+        GLOB_LOGGER("general");
+        BOOST_LOG_SEV(logger, Info) << " Unexpected call Oo ! " << ec.message();
+    }
 }
         
         
 void delayed(ActiveObjectPtr) {
+    
+    GLOB_LOGGER("general");
+
+    BOOST_LOG_SEV(logger, Info) << " Setting up Tests workflow !";
     // timeout execution.
-    Timed t;
-    t.setIOService(service);
-    t.setDuration(1000);
-    t.setTimeoutFunction(&terminate);
-    t.start();
+    t.reset(new Timed());
+    t->setIOService(service);
+    t->setDuration(5000);
+    t->setTimeoutFunction(&terminate);
+    t->start();
     
     WorkflowPtr workflow(new Workflow("test-workflow-a"));
     auto sm = workflow->getStateMachine();
@@ -62,6 +75,8 @@ int main(int argc, const char * argv[]) {
     // maybe we should wait for controller to have started ...
     service.reset(new boost::asio::io_service());
     Logged::loadConfiguration("");
+    GLOB_LOGGER("general");
+    BOOST_LOG_SEV(logger, Info) << " Test Starting !";
     ControllerManager::getInstance()->setStartedFunction(&delayed);
     ControllerManager::getInstance()->start();
